re-organize splash and intro folders
