+++ lib/features/splash/view/widgets/splash_background.dart
