+++ lib/features/splash/view/widgets/splash_icon.dart
